+++ CHANGELOG.md
@@ -6,10 +6,13 @@ Notable changes will be documented in this file. The project adheres to [Semanti
 Unreleased
 ----------
 
+* Improved `AbstractRepository`
+  * `findByPid` returns a query result instead of the first object
 * Improved `ArrayViewHelper`
-* Added `crdate` to `AbstractEntity`
+* Added `crdate` property to `AbstractEntity`
 * Use `call_user_func` in `ext_localconf.php` and `ext_tables.php`
 * Updated comments
+* Cleaned up
 
 1.8.0
 -----